refactor: 5단계 리팩토링
