+++ src/main/java/racinggame/view/ResultView.java
@@ -2,7 +2,7 @@ package racinggame.view;
 
 import racinggame.common.constants.RexFormatConstants;
 import racinggame.common.util.PrintUtils;
-import racinggame.vo.Car;
+import racinggame.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,7 +13,6 @@ public class ResultView {
     private static final String RESULT_TEXT = "실행 결과";
     private static final String RESULT_WINNER_TEXT = "가 최종 우승했습니다.";
     private static final StringBuffer sb = new StringBuffer();
-    private static final String LAST_COMMA_REMOVE_REX_FORMAT_TEXT = ",$";
     private static int winnerMoveCount = 0;
     private static List<Car> winnerCarList = new ArrayList<>();
 
@@ -44,9 +43,9 @@ public class ResultView {
         }
         StringBuilder sb = new StringBuilder();
         for (Car car : winnerCarList) {
-            sb.append(car.getName()).append(RexFormatConstants.CARS_NAME_REX_FORMAT_TEXT.getValue());
+            sb.append(car.getName()).append(RexFormatConstants.CARS_NAME_REX_FORMAT_TEXT.getFormat());
         }
-        return sb.toString().replaceAll(LAST_COMMA_REMOVE_REX_FORMAT_TEXT, "");
+        return sb.toString().replaceAll(RexFormatConstants.LAST_COMMA_REMOVE_REX_FORMAT_TEXT.getFormat(), "");
     }
 
     private static void addWinner(Car car) {
