+++ src/test/java/racinggame/RacingCarGameTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racinggame.vo.Car;
+import racinggame.domain.Car;
 
 import java.util.Map;
